created slug homepage
